+++ tests/testthat/test-checkArgs.R
@@ -12,34 +12,35 @@ contrasts <- c(1, 2)
 penalties <- c(1, 2)
 clust_method <- "kmeans"
 linkage_method <- "complete"
+clusters <- NULL
 
 test_that("Only data.frames, tibbles, matrices, and sparse matrices pass", {
   expect_silent(checkArgs(
     toy_df[, 1:30], background_df, center,
     scale, n_eigen, contrasts, penalties,
-    clust_method, linkage_method
+    clust_method, linkage_method, clusters
   ))
   expect_silent(checkArgs(
     as.matrix(toy_df[, 1:30]), as.matrix(background_df),
     center, scale, n_eigen, contrasts, penalties,
-    clust_method, linkage_method
+    clust_method, linkage_method, clusters
   ))
   expect_silent(checkArgs(
     as_tibble(toy_df[, 1:30]), as_tibble(background_df),
     center, scale, n_eigen, contrasts, penalties,
-    clust_method, linkage_method
+    clust_method, linkage_method, clusters
   ))
   expect_silent(checkArgs(
     as(as.matrix(toy_df[, 1:30]), "dgCMatrix"),
     as(as.matrix(background_df), "dgCMatrix"),
     center, scale, n_eigen, contrasts, penalties,
-    clust_method, linkage_method
+    clust_method, linkage_method, clusters
   ))
   expect_silent(checkArgs(
     as(as.matrix(toy_df[, 1:30]), "dgeMatrix"),
     as(as.matrix(background_df), "dgeMatrix"),
     center, scale, n_eigen, contrasts, penalties,
-    clust_method, linkage_method
+    clust_method, linkage_method, clusters
   ))
 })
 
@@ -50,7 +51,7 @@ test_that(paste(
   expect_error(checkArgs(
     toy_df, background_df, center,
     scale, n_eigen, contrasts, penalties,
-    clust_method, linkage_method
+    clust_method, linkage_method, clusters
   ),
   "ncol(target) not equal to ncol(background)",
   fixed = TRUE
@@ -61,42 +62,42 @@ test_that("Center and scale arguments only handle Logical options", {
   expect_silent(checkArgs(
     toy_df[, 1:30], background_df, center,
     scale, n_eigen, contrasts, penalties,
-    clust_method, linkage_method
+    clust_method, linkage_method,clusters
   ))
   expect_silent(checkArgs(
     toy_df[, 1:30], background_df, FALSE,
     TRUE, n_eigen, contrasts, penalties,
-    clust_method, linkage_method
+    clust_method, linkage_method, clusters
   ))
   expect_silent(checkArgs(
     toy_df[, 1:30], background_df, TRUE,
     FALSE, n_eigen, contrasts, penalties,
-    clust_method, linkage_method
+    clust_method, linkage_metho, clusters
   ))
   expect_silent(checkArgs(
     toy_df[, 1:30], background_df, FALSE,
     FALSE, n_eigen, contrasts, penalties,
-    clust_method, linkage_method
+    clust_method, linkage_method, clusters
   ))
   expect_error(checkArgs(
     toy_df[, 1:30], background_df, 1,
     scale, n_eigen, contrasts, penalties,
-    clust_method, linkage_method
+    clust_method, linkage_method, clusters
   ))
   expect_error(checkArgs(
     toy_df[, 1:30], background_df, "a",
     scale, n_eigen, contrasts, penalties,
-    clust_method, linkage_method
+    clust_method, linkage_method, clusters
   ))
   expect_error(checkArgs(
     toy_df[, 1:30], background_df, center,
     "scale", n_eigen, contrasts, penalties,
-    clust_method, linkage_method
+    clust_method, linkage_method, clusters
   ))
   expect_error(checkArgs(
     toy_df[, 1:30], background_df, center,
     12342, n_eigen, contrasts, penalties,
-    clust_method, linkage_method
+    clust_method, linkage_method, clusters
   ))
 })
 
@@ -104,32 +105,32 @@ test_that("Argument n_eigen is set to an integer between 1 and ncol(target)", {
   expect_silent(checkArgs(
     toy_df[, 1:30], background_df, center,
     scale, 1, contrasts, penalties,
-    clust_method, linkage_method
+    clust_method, linkage_method, clusters
   ))
   expect_silent(checkArgs(
     toy_df[, 1:30], background_df, center,
     scale, 30, contrasts, penalties,
-    clust_method, linkage_method
+    clust_method, linkage_method, clusters
   ))
   expect_error(checkArgs(
     toy_df[, 1:30], background_df, center,
     scale, 31, contrasts, penalties,
-    clust_method, linkage_method
+    clust_method, linkage_method, clusters
   ))
   expect_error(checkArgs(
     toy_df[, 1:30], background_df, center,
     scale, 0, contrasts, penalties,
-    clust_method, linkage_method
+    clust_method, linkage_method, clusters
   ))
   expect_error(checkArgs(
     toy_df[, 1:30], background_df, center,
     scale, "n_eigen", contrasts, penalties,
-    clust_method, linkage_method
+    clust_method, linkage_method, clusters
   ))
   expect_error(checkArgs(
     toy_df[, 1:30], background_df, center,
     scale, 1.5, contrasts, penalties,
-    clust_method, linkage_method
+    clust_method, linkage_method, clusters
   ))
 })
 
@@ -137,27 +138,27 @@ test_that("Contrasts is a non-negative, non-zero vector of length >= 1", {
   expect_silent(checkArgs(
     toy_df[, 1:30], background_df, center,
     scale, n_eigen, 1, penalties,
-    clust_method, linkage_method
+    clust_method, linkage_method, clusters
   ))
   expect_silent(checkArgs(
     toy_df[, 1:30], background_df, center,
     scale, n_eigen, seq(1, 10, by = 0.1), penalties,
-    clust_method, linkage_method
+    clust_method, linkage_method, clusters
   ))
   expect_error(checkArgs(
     toy_df[, 1:30], background_df, center,
     scale, n_eigen, 0, penalties,
-    clust_method, linkage_method
+    clust_method, linkage_method, clusters
   ))
   expect_error(checkArgs(
     toy_df[, 1:30], background_df, center,
     scale, n_eigen, c(seq(1, 10, by = 0.1), -1),
-    penalties, clust_method, linkage_method
+    penalties, clust_method, linkage_method, clusters
   ))
   expect_error(checkArgs(
     toy_df[, 1:30], background_df, center,
     scale, n_eigen, "contrasts", penalties,
-    clust_method, linkage_method
+    clust_method, linkage_method, clusters
   ))
 })
 
@@ -165,28 +166,28 @@ test_that("Penalties is a non-negative vector of length at least 1", {
   expect_silent(checkArgs(
     toy_df[, 1:30], background_df, center,
     scale, n_eigen, contrasts, penalties,
-    clust_method, linkage_method
+    clust_method, linkage_method, clusters
   ))
   expect_silent(checkArgs(
     toy_df[, 1:30], background_df, center,
     scale, n_eigen, contrasts, seq(1, 10, by = 0.1),
-    clust_method, linkage_method
+    clust_method, linkage_method, clusters
   ))
   expect_silent(checkArgs(
     toy_df[, 1:30], background_df, center,
     scale, n_eigen, contrasts, 0,
-    clust_method, linkage_method
+    clust_method, linkage_method, clusters
   ))
   expect_error(checkArgs(
     toy_df[, 1:30], background_df, center,
     scale, n_eigen, contrasts,
     c(seq(1, 10, by = 0.1), -1),
-    clust_method, linkage_method
+    clust_method, linkage_method, clusters
   ))
   expect_error(checkArgs(
     toy_df[, 1:30], background_df, center,
     scale, n_eigen, contrasts, "penalties",
-    clust_method, linkage_method
+    clust_method, linkage_method, clusters
   ))
 })
 
@@ -194,26 +195,49 @@ test_that("ward.D linkage method cannot be used when clust_method is hclust", {
   expect_silent(checkArgs(
     toy_df[, 1:30], background_df, center,
     scale, n_eigen, contrasts, penalties,
-    clust_method, linkage_method
+    clust_method, linkage_method, clusters
   ))
   expect_silent(checkArgs(
     toy_df[, 1:30], background_df, center,
     scale, n_eigen, contrasts, penalties,
-    clust_method = "pam", linkage_method
+    clust_method = "pam", linkage_method, clusters
   ))
   expect_silent(checkArgs(
     toy_df[, 1:30], background_df, center,
     scale, n_eigen, contrasts, penalties,
-    clust_method = "hclust", linkage_method = "single"
+    clust_method = "hclust", linkage_method = "single", clusters
   ))
   expect_silent(checkArgs(
     toy_df[, 1:30], background_df, center,
     scale, n_eigen, contrasts, penalties,
-    clust_method = "hclust", linkage_method = "ward.D2"
+    clust_method = "hclust", linkage_method = "ward.D2", clusters
   ))
   expect_error(checkArgs(
     toy_df[, 1:30], background_df, center,
     scale, n_eigen, contrasts, penalties,
-    clust_method = "hclust", linkage_method = "ward.D"
+    clust_method = "hclust", linkage_method = "ward.D", clusters
+  ))
+})
+
+test_that("Catches cluster assignments that don't match requirements", {
+  expect_silent(checkArgs(
+    toy_df[, 1:30], background_df, center,
+    scale, n_eigen, contrasts, penalties,
+    clust_method, linkage_method , clusters
+  ))
+  expect_silent(checkArgs(
+    toy_df[, 1:30], background_df, center,
+    scale, n_eigen, contrasts, penalties,
+    clust_method, linkage_method , clusters = toy_df[, 31]
+  ))
+  expect_error(checkArgs(
+    toy_df[, 1:30], background_df, center,
+    scale, n_eigen, contrasts, penalties,
+    clust_method, linkage_method , clusters = toy_df[1:25, 31]
+  ))
+  expect_error(checkArgs(
+    toy_df[, 1:30], background_df, center,
+    scale, n_eigen, contrasts, penalties,
+    clust_method, linkage_method , clusters = as.charaacter(toy_df[, 31])
   ))
 })